+++ client/src/components/manager/GameManager.jsx
