+++ client/src/components/manager/game_manager.css
